Complete Weight/Rig runtime decomposition
