+++ src/tests/web/test_app_flow.py
@@ -70,9 +70,6 @@ def test_frontend_public_surface_and_lifecycle_events(edge_browser, frontend_url
             "getAmbientOcclusionStrength", "getBloomEnabled", "getCurrentSource",
             "getEnvironmentPreset", "getLoadBenchmark", "getMaterialState",
             "getOutlineState", "getRenderCount", "openMod",
-            "getModelPhysicsState",
-            "getModelRigState",
-            "getModelWeightState", "setModelWeightHeatmap",
             "refreshControlSemantics", "refreshMeshSemantics",
             "refreshPresentState", "reloadCurrentMod", "setEnvironmentPreset",
             "setBloomEnabled", "setMaterialDebugMode", "setOutlineEnabled", "switchAsset",
